Support hardware 3level stack for small AVR devices.
Some small AVR devices such as the venerable AT90S1200 or the
newer ATTiny15L do not have RAM and thus do not have a RAM based
stack. The new class ThreeLevelStack, derived from MemoryOffsets
implements the stack area for a three level stack.
It also changes the semantics of the stack a bit, such that
a stack ceiling and not a stack mask is specified. This allows
for stack sizes which are not a power of two.