Finishing the pluggable C++ factory for AVR devices
This implements a decoupling of particular devices from any
code which uses them through the interface as it is presented
in avrfactory.h. It should make it easier to support various
external interfaces to the simulavrxx code.

One needs to compile simulavrxx with dynamic linking, though.
Without dynamic linking the factory code will not work (at least
not out of the box).