Configurable bit fields for timer 01 IRQ bits
Allow configuration of the timer 01 IRQ bit field positions by
the initialization code of an AVR device implementation.

For example, in case of the Mega8, the bit fields are
different (like for the mega0123 timer), but the Mega8 has
only three timers (not four). Changing just the bit fields
allows to simulate the timer1 behaviour of the mega8 for
moderately complex situations.

Additionally, the hard coded Tifr flag values have been replaced.