AvrDevice::ReplaceIoRegister() should accept correct I/O range
This fixes the allowed range in AvrDevice::ReplaceIoRegister which
was errorneously checked against too strict limits.

TODO: Because ReplaceIoRegister gets a offset into the AVR data address range,
it should maybe even renamed to somewhat more sensible (the current name is
misleading as it is unclear whether the offset should be added before or not.