Rewrites a former patch for using libbfd.so instead of libbfd.a, if found
The former patch "This BFD linking with the shared object libbfd.so works much better" from Onno
has made a problem, if libbfd.a was found, but not libbfd.so, it has assumed allways, that
libbfd.so was available. Now it checks existence of both variants and tries to use libbfd.so first
and then libbfd.a. Same is introduced for libiberty.so/libiberty.a. Furter, linking against this
libs on src/Makefile.am is made explicit. It uses the right libpath and let linker decide, what's
the right lib to link against.