Only build avr-fab devices when the XML files are available
This adds a missing check to configure (and src/Makefile.am) that only enable
building of the automatically produced avr-devices when the XML part
description files have been placed into the src/fab/atmel-xml directory.

As it only enables the complicated build process (which depends on other
software such as python cheetah) when one willingly places these files into
this directory, is at least now ensured that no strange error messages occur
for those who simply want to compile the bare bones of simulavrxx.

In the future, this functionality should either be amended to include checks
for the necessary python infrastructure or the files generated from the XML
part description files should be added to the repository and the tools to
built them will be supplied to regenerate them for those who have installed
the XML files. (As suggested by Thomas Klepp)