Workaround for a major memory leak in verilog mode
The MirrorNets created by the Pins will not be deleted when a Pin gets
destroyed; this will lead to a major memory leak in verilog VPI interface mode
which makes this interface unusable for longer runs. This patch fixes this
memory leak in VPI mode, but should not touch the memory management of
simulavrxx in normal mode, as other bugs in the memory management will appear
when doing so.

TODO: This is only a partial workaround and should replaced ASAP with a proper
fix of the various memory leaks in simulavrxx.