Simple SPI master waveform output check in c/verilog
This small file produces a GTKwave-viewable .vcd file which can be used to
check the default waveforms of the simulated SPI interface in simulavrxx, at
the moment only in master mode. It has been written to aid in debugging the
SPI interface.

TODO: It should probably be extended to cover more stuff of the SPI interface
and maybe integrated into a future regression test suite.