Tracing infrastructure and rework of the I/O register system
This quite large EXPERIMENTAL patch modifies the register system so that all
registers can be linked to a tracing object which can then be activated for
certain registers or values and all accesses to such instrumented registers
can be written to file using a generic 'trace output writer' (dumper)
mechanism. Two new options, '-o' and '-c' have been implemented in main.cpp to
be able to use this new functionality from the command line.

The I/O register rework removes a lot of redundant code and simplifies the I/O
register infrastructure by reducing all registers in hardware components to
instantiations of a single IOReg<T> template class which has to be linked to
the getter/setter methods inside the hardware class. It also concentrates the
data flow of register set/get operations into the two virtual methods
RWMemoryMember::get() and RWMemoryMember::set(). This change is then used to
hook the tracing infrastructure into the register access methods.

Tracing for RWMemoryMembers is done in two subsequent steps, per CPU cycle:
1. Any access to a register with enable tracing will be logged, saving
different access modes etc.
2. At the end of a CPU cycle touching several registers, the changes are
collected from the TraceValue instances and a Dumper instances is fed the
changes which will in turn write and output file or collect statistics.

A special exception are traced values which do not correspond to a
RWMemoryMember; their underlying values are simply checked for changes during
each cycle.

Two simple tracing outputs have been implemented so far:
--------------------------------------------------------
WarnUnknown: A very simple trace dumper which will warn when a memory value is
read accessed but has not been written before.

DumpVCD: Produces .vcd value change dump files which can be viewed with an
appropiate viewer, for example with gtkwave. In addition to the values
themselves, it can also trace the read and write accesses to registers using
virtual strobe signal lines.

Performance
-----------
Registers which are not enabled for tracing are just doing a check of a
boolean value per access. As the old tracing code (using the (core->trace_on==1)
check), the performance SHOULD be similar, at least when tracing is disabled.

TODO
----
- Integrate all old tracing code still lying around / remove redundant stuff
- Thorough documentation of the tracing infrastructure is still missing, as well
as any development documentation.
- More sophisticated tracers could be implemented.
- Testing, testing, testing!
- Implement shadow register tracing for more hardware. Currently, it is only
(partly) implemented for SPI and the U(S)ART