Make building API doc with doxygen depending on configure switch and existence of doxygen
To build API documentation with doxygen there is now a switch for ./configure: --enable-doxygen-doc.
If switched on and doxygen is found in path, then you can call "make doxygen-doc" in root directory
to build API documentation. This documentation wil be build in doc/doxygen-doc directory. Even it's
possible to create other formats, only html is enabled by default.