Rewrite of pysimulavr.i to get it working
This is the start of rewriting pysimulavr.i to get python interface real functional.

Attention! There is also a change on src/Makefile.am. It's necessary to keep the
order of linking: libbfd before libiberty. There was a case (in Ubuntu 9.04), where
libbfd.a was available, but not libbfd.so. With the wrong order it was linked against
/usr/lib/libbfd.so, because libiberty.so was located there. Accessing functions, which
have used libbfd have caused then a segfault!