Extend HWTimerPrescaler class
This commit is the first step for a rework of timer hardware units. Prescaler get reset functionality,
for this a IOSpecialReg class is introduced. IOSpecialReg is used for control register, which are not
unique for one and only one hardware unit (as example GTCCR register for ATMega48). A hardware modul can
be connected to this register to get informed on read or write to this register. Prescaler value can
now also traced by DumpManager infrastructure.