Bugfix: DumpVCD wrotes incomplete files
DumpVCD has wrote incomplete VCD-files if simulavr is used by comandline. Reason was, that for
performance DumpVCD hasn't uses implicit flushing of file by using endl. But in main the device
wasn't deleted after simulation, so a flush on closing file wasn't performed. Further there is
a improvement in DumpVCD, not to write time marker, if nothing is changed. This avoids writing
of big VCD files, if it isn't necessary.