Introduce TraceValueRegister class to get a possibility to build a name hierarchy for TraceValue names
This class will replace the big all-map in DumpManager. Further it will give possibility to order
trace values in case of more than one device. In the moment AVRDevice has no name for TraceValue
names, but TraceValueRegister has a contructor to hold the name (for example "dev1", dev2" ...)
in case of simulation with more than one device.