Make Makefile.am from Makefile in examples/verilog
To call examples goto examples/verilog and type "make" or "make examples". This works
only, if you have verilog configured, if you have cross compilation environment and
verilog tools like iverilog and vvp and if you have xml files for attiny2313 and atmega8
in the fab directory. And, of course, if make was successfull!