Rewrite avrerror.cpp: adapt to c++ and add some new features
The default behaviour is the same as before. But now it's possible, to
raise a exception instead of using exit on calling avr_error. This is
needed for example for python interface to catch such errors instead of
hard breaking the application. But it could also used for other things.
A possibility to redirect message, warning and error output to a stream
is also added. So it's possible to catch such output too in a simple way.