Eliminate usage of exit call (except main.cpp), use functions in avrerror instead
Centralises all exit points to one place, except main.cpp, to get a common
handling on exit. Now all errors exits return 1 on simulavr (and if not
switched to raise a exeption) printable.cpp is removed, because it's better
to solve, if Printable is changed to be a abstract class.