Fix of template files for AVR factory for newer version of python-cheetah
At least on Ubuntu 9.10, the default-installed python-cheetah, version
2.0.1-2ubuntu6, will not any longer accept '#endfor' or '#endif' for
closing '#for' resp. '#if' constructs. Rather, it needs to have an
extra space in between.

This is only the first in a series of patches to get the factory working
again, after it suffered bitrot from the newer set of changes and
enhancements. A version with working factory should be in commit
8800f75deebd43026fa802f1cc99df2d99b49c30.