Fix for the dependencies to remake avr.vpi properly
If any of the files for libsim_la changes, the avr.vpi file now gets remade
properly. Adding just the objects to the LIBADD was not enough as this will
not generate the proper dependencies (on all the source files) for
avr.vpi. Now, the library avr.vpi is simply remade by depending on all
libsim-objects.