Export of SetDumpTraceArgs as $avr_dumparg() to Verilog
This exports the SetDumpTraceArgs trace value argument processor function to
verilog. Note that only one string can be given per time and not a vector of
strings, as it is the case for SetDumpTraceArgs. As there is only a
simple iteration over all argument strings, this should not be a problem.

Please note that in Makefile.am, it was necessary to resolve a libtool issue
with cmd/dumpargs.cpp being built twice. For more documentation, see

http://www.gnu.org/software/libtool/manual/automake/Objects-created-both-with-libtool-and-without.html