AVR VPI interface: start, stop and set-time implementation
This implements starting and stopping of the VCD dumping process as well as
setting the current time in the AVR core to the value as given by the Verilog
simulation system. With this patch, VCD files can be generated from within the
AVR simulator core as well as from the verilog simulator. To view both of them
in phase, use a tool such as 'twinview'.