Fix program counter so that it stays on the 'right PC' for long commands
The program counter values were off, both when used from the Verilog interface
and in the 'usual tracing output' as well. This should fix the program counter
displays by introducing an additional cPC variable that will be kept at the
same value for the whole duration of the execution of an assembly instruction.