Add BREAK-instruction, causes simulavr to halt.
Atmel instruction-set datasheet documents instruction "BREAK" which
behaves like a breakpoint on real chip if JTAG/debugWIRE adapter is
connected and debugging fuse is enabled and the chip supports the
instruction.

TODO: We could write a message if the instruction is not supported
by a chip, and break as usual. The feature is useful for assertions.

Many thanks to Stan Behrens for the patch.
http://lists.nongnu.org/archive/html/simulavr-devel/2012-01/msg00031.html