Add yet another example, simple_serial.
This example is extremely simple, it does nothing fancy, breath taking
or showing advanced coding skills. It also depends on neither TCL nor
Python. It's purpose is to exercise the UART serial port and to show
how to make this port available in a simulation _without_ altering the
binary.

The code is tested on real hardware already and works just as expected.